requirements.txt
tensorflow
Pillow 
numpy
streamlit